Merging in updated code from Misha and Sergey - Release v1

+++ gemdaq-testing/gemsupervisor/include/gem/supervisor/GEMGLIBSupervisorWeb.h
@@ -237,7 +237,7 @@ namespace gem {
         std::shared_ptr<gem::readout::GEMDataParker> gemDataParker;
 
         // Counter
-        int counter_[3];
+        uint64_t counter_[3];
 
         // VFAT Blocks Counter
         int vfat_;
